chore: allow changing browser provider per workspace
